refactor: move user manager into internal/core
