+++ internal/core/approvals/userClaims_test.TestClaims.example.output.nauth.approved.yaml
