+++ internal/core/approvals/userClaims_test.TestClaims.permissions.output.nats.approved.json
