+++ internal/core/approvals/userClaims_test.TestClaims.permissions.output.nauth.approved.yaml
