+++ internal/core/user_test.go
@@ -1,4 +1,4 @@
-package user
+package core
 
 import (
 	"context"
