--- internal/user/mocks_test.go
@@ -1,110 +1,0 @@
-package user
-
-import (
-	"context"
-
-	"github.com/WirelessCar/nauth/internal/domain"
-	"github.com/WirelessCar/nauth/internal/ports"
-	"github.com/nats-io/jwt/v2"
-	"github.com/stretchr/testify/mock"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-/* ****************************************************
-* JWTSigner Mock
-*****************************************************/
-
-func NewUserJWTSignerMock() *UserJWTSignerMock {
-	return &UserJWTSignerMock{}
-}
-
-type UserJWTSignerMock struct {
-	mock.Mock
-}
-
-func (m *UserJWTSignerMock) SignUserJWT(ctx context.Context, accountRef domain.NamespacedName, claims *jwt.UserClaims) (*SignedUserJWT, error) {
-	args := m.Called(ctx, accountRef, claims)
-	return args.Get(0).(*SignedUserJWT), args.Error(1)
-}
-
-func (m *UserJWTSignerMock) mockSignUserJWT(ctx context.Context, accountRef domain.NamespacedName, callback func(claims *jwt.UserClaims) *SignedUserJWT) {
-	call := m.On("SignUserJWT", ctx, accountRef, mock.Anything)
-	call.RunFn = func(args mock.Arguments) {
-		claims := args.Get(2).(*jwt.UserClaims)
-		call.Return(callback(claims), nil)
-	}
-}
-
-var _ JWTSigner = &UserJWTSignerMock{}
-
-/* ****************************************************
-* ports.SecretClient Mock
-*****************************************************/
-
-func NewSecretClientMock() *SecretClientMock {
-	return &SecretClientMock{}
-}
-
-type SecretClientMock struct {
-	mock.Mock
-}
-
-// Apply implements ports.SecretStorer.
-func (s *SecretClientMock) Apply(ctx context.Context, owner metav1.Object, meta metav1.ObjectMeta, valueMap map[string]string) error {
-	args := s.Called(ctx, owner, meta, valueMap)
-	return args.Error(0)
-}
-
-func (s *SecretClientMock) mockApply(ctx context.Context, owner interface{}, meta interface{}, valueMap interface{}) {
-	s.On("Apply", ctx, owner, meta, valueMap).Return(nil)
-}
-
-func (s *SecretClientMock) mockApplyWithCatch(ctx context.Context, owner interface{}, meta interface{}, valueMap interface{}, catcher func(map[string]string)) {
-	s.On("Apply", ctx, owner, meta, valueMap).
-		Run(func(args mock.Arguments) {
-			catcher(args.Get(3).(map[string]string))
-		}).
-		Return(nil)
-}
-
-// Get implements ports.SecretStorer.
-func (s *SecretClientMock) Get(ctx context.Context, secretRef domain.NamespacedName) (map[string]string, error) {
-	args := s.Called(ctx, secretRef)
-	return args.Get(0).(map[string]string), args.Error(1)
-}
-
-// GetByLabels implements ports.SecretStorer.
-func (s *SecretClientMock) GetByLabels(ctx context.Context, namespace domain.Namespace, labels map[string]string) (*corev1.SecretList, error) {
-	args := s.Called(ctx, namespace, labels)
-	return args.Get(0).(*corev1.SecretList), args.Error(1)
-}
-
-// DeleteSecret implements ports.SecretStorer.
-func (s *SecretClientMock) Delete(ctx context.Context, secretRef domain.NamespacedName) error {
-	args := s.Called(ctx, secretRef)
-	return args.Error(0)
-}
-
-func (s *SecretClientMock) mockDelete(ctx context.Context, secretRef domain.NamespacedName) {
-	s.On("Delete", ctx, secretRef).Return(nil)
-}
-
-func (s *SecretClientMock) mockDeleteError(ctx context.Context, secretRef domain.NamespacedName, err error) {
-	s.On("Delete", ctx, secretRef).Return(err)
-}
-
-// DeleteSecret implements ports.SecretStorer.
-func (s *SecretClientMock) DeleteByLabels(ctx context.Context, namespace domain.Namespace, labels map[string]string) error {
-	args := s.Called(ctx, namespace, labels)
-	return args.Error(0)
-}
-
-// LabelSecret implements ports.SecretStorer.
-func (s *SecretClientMock) Label(ctx context.Context, secretRef domain.NamespacedName, labels map[string]string) error {
-	args := s.Called(ctx, secretRef, labels)
-	return args.Error(0)
-}
-
-// Compile-time assertion that implementation satisfies the ports interface
-var _ ports.SecretClient = &SecretClientMock{}
