--- internal/user/predicates_test.go
@@ -1,11 +1,0 @@
-package user
-
-import (
-	"strings"
-)
-
-var (
-	isUserPubKey = func(input string) bool {
-		return strings.HasPrefix(input, "U")
-	}
-)
